Split out formatting task from external formatter
